2022-17 part 2: alright, I cheated...
So, I found this blog online that had two brilliant ideas (one of which
I kinda had, but didn't know how to capitalize on, more on that later).

1. Rocks can be represented as 32bit values, 8bits per row

   That means we get to do some bitwise operations to check for
   collisions!
2. Surely there's a cycle to be found somewhere!

   This is the one I'd already surmised, but didn't know what to do
   about. Since the rocks and the jets are on a cycle, there's bound to
   be a cycle somewhere in where rocks settle in the well. I just didn't
   know what the next variable ought to be. Turns out that it's the
   canopy: the distance from the top of the topmost settled rock in each
   column.

   By stuffing those three as a key into a HashMap, with the value being
   the count of rocks already dropped and the height of the well, we can
   stop once we've found a cycle, the do some simple additions.

Anyway, to not be a complete cheat, I did reimplement most myself, and
tweaked a few things (the actual solution in the blog I found didn't
work for me, as it didn't use the canopy, just the top 8 rows of the
well, which wasn't enough for my input).