2023-15: Part 2
I'm using the `itermap` crate here, for a map that maintains insertion
order (which we need). I'm contemplating doing it myself by making boxes
a list of vecs, but then I'd need to iterate over them everytime.

So I'll have to benchmark it. On my M1 Pro, the current implementation
takes about 2ms.