2023-05: Part two finally fast...
I can barely understand what I wrote there. But the gist of it is: map
entire ranges at once. To do this, you map whatever overlap a range has
with a map, and try to map the leftovers with the next map, etc.

There's a terminology issue: it's maps all the way down!