2024-17 Part 1 done
Dammit, that was painful. Implementing all that wasn't too hard, but I'd
made the assumption that jumps would always land on what was originally
an instruction. But you could definitely jump to let's say 1, which
means that whatever was in `instructions[1]` is no longer an operand,
but now an instruction!

So that's why operands _must_ have the same value range as
instructions... Me dumdum!