2024-18 Done!
I have no idea why, but using a BinaryHeap as a priority queue (even
when using Reverse to make it a min-heap) results in an infinite loop
(or at least a very long running loop).
Maybe it has something to do with me not doing anything to prevent
tracking back (that's probably it).

Oh, and Grid has a new `with_size` constructor. It requires T to
implement Default, which is easy enough in most occasions =)