2024-21 Kinda done...
I have to run the solution a couple of times, and wait for the smallest
answer. Something about my heuristic for picking the "shortest path"
from one key to the next isn't good enough. I just now thought of
another thing to add though: keys that are further away from the
confirmation button `A` should preferably be at the end of the path,
especially if they repeat, as after two expansions, it'll explode?