2023-23 Parts 1 & 2 done
Part 2 is dog slooooow. Time to read up on algorithms!
Also, I can't believe part 2 works. Instead of taking the adjacency list
in the input, and making sure to consider each adjacency bidirectional
(as we're dealing with an undirected graph), I'm only looking in the
direction as given in the input. How this works is beyond me, probably
an artifact of the way the inputs are generated.