2019-07: Part 2 done!
I had some silly problems, such as forgetting that I don't need to alter
the instruction pointer when halting a Computer...

Also added a `SplitIO` struct, that has both an input and an output,
separate from one another. To use it with some VecDeque's I had in a
Vec, I had to use Vec::get_disjoint_mut, which is interesting!